# docs/api/pipeline-openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRetrievalProfileSchema": {
        "additionalProperties": false,
        "description": "Register one retrieval profile under the server-controlled consumer scope.",
        "properties": {
          "corpus_version": {
            "minLength": 1,
            "title": "Corpus Version",
            "type": "string"
          },
          "embedding_profile_id": {
            "minLength": 1,
            "title": "Embedding Profile Id",
            "type": "string"
          },
          "indexing_target_id": {
            "minLength": 1,
            "title": "Indexing Target Id",
            "type": "string"
          },
          "lexical_fallback_policy": {
            "default": "allowed_when_vector_unavailable",
            "title": "Lexical Fallback Policy",
            "type": "string"
          }
        },
        "required": [
          "corpus_version",
          "embedding_profile_id",
          "indexing_target_id"
        ],
        "title": "CreateRetrievalProfileSchema",
        "type": "object"
      },
...
      "ErrorBodySchema": {
        "additionalProperties": false,
        "description": "Body of the shared error envelope.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBodySchema",
        "type": "object"
      },
      "ErrorEnvelopeSchema": {
        "additionalProperties": false,
        "description": "Every non-2xx response uses this envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodySchema"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelopeSchema",
        "type": "object"
      },
...
      "PaginatedRetrievalProfilesSchema": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RetrievalProfileSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total_items": {
            "minimum": 0.0,
            "title": "Total Items",
            "type": "integer"
          },
          "total_pages": {
            "minimum": 0.0,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total_items",
          "total_pages"
        ],
        "title": "PaginatedRetrievalProfilesSchema",
        "type": "object"
      },
...
      "RetrievalProfileSchema": {
        "additionalProperties": false,
        "description": "Read-only view of one retrieval profile.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "consumer_scope_id": {
            "title": "Consumer Scope Id",
            "type": "string"
          },
          "consumer_scope_type": {
            "title": "Consumer Scope Type",
            "type": "string"
          },
          "corpus_version": {
            "title": "Corpus Version",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "deprecated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deprecated At"
          },
          "embedding_profile_id": {
            "title": "Embedding Profile Id",
            "type": "string"
          },
          "indexing_target_id": {
            "title": "Indexing Target Id",
            "type": "string"
          },
          "last_runtime_status": {
            "title": "Last Runtime Status",
            "type": "string"
          },
          "lexical_fallback_policy": {
            "title": "Lexical Fallback Policy",
            "type": "string"
          },
          "retrieval_profile_id": {
            "title": "Retrieval Profile Id",
            "type": "string"
          },
          "validated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated At"
          },
          "validation_status": {
            "title": "Validation Status",
            "type": "string"
          }
        },
        "required": [
          "retrieval_profile_id",
          "consumer_scope_type",
          "consumer_scope_id",
          "corpus_version",
          "embedding_profile_id",
          "indexing_target_id",
          "lexical_fallback_policy",
          "active",
          "validation_status",
          "last_runtime_status"
        ],
        "title": "RetrievalProfileSchema",
        "type": "object"
      },
      "RetrievalProfileStatusSchema": {
        "additionalProperties": false,
        "description": "Profile, runtime and readiness in one payload.",
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/RetrievalProfileSchema"
          },
          "readiness": {
            "$ref": "#/components/schemas/RetrievalReadinessSchema"
          },
          "runtime": {
            "$ref": "#/components/schemas/RetrievalRuntimeSchema"
          }
        },
        "required": [
          "profile",
          "runtime",
          "readiness"
        ],
        "title": "RetrievalProfileStatusSchema",
        "type": "object"
      },
      "RetrievalReadinessSchema": {
        "additionalProperties": false,
        "description": "Every precondition retrieval needs, evaluated together.",
        "properties": {
          "active_document_count": {
            "minimum": 0.0,
            "title": "Active Document Count",
            "type": "integer"
          },
          "active_vector_rows": {
            "minimum": 0.0,
            "title": "Active Vector Rows",
            "type": "integer"
          },
          "blocking_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Blocking Reasons",
            "type": "array"
          },
          "embedding_bundle_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Bundle Id"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "retrieval_profile_id": {
            "title": "Retrieval Profile Id",
            "type": "string"
          }
        },
        "required": [
          "retrieval_profile_id",
          "ready",
          "active_vector_rows",
          "active_document_count"
        ],
        "title": "RetrievalReadinessSchema",
        "type": "object"
      },
...
      "RetrievalSearchResponseSchema": {
        "additionalProperties": false,
        "description": "Evidence returned by one retrieval search.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RetrievedEvidenceSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "retrieval_profile_id": {
            "title": "Retrieval Profile Id",
            "type": "string"
          },
          "top_k": {
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "retrieval_profile_id",
          "top_k"
        ],
        "title": "RetrievalSearchResponseSchema",
        "type": "object"
      },
      "RetrievalValidationSchema": {
        "additionalProperties": false,
        "description": "Result of a controlled smoke validation.",
        "properties": {
          "blocking_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Blocking Reasons",
            "type": "array"
          },
          "candidates_found": {
            "minimum": 0.0,
            "title": "Candidates Found",
            "type": "integer"
          },
          "query_dimension": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Dimension"
          },
          "retrieval_profile_id": {
            "title": "Retrieval Profile Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "validator_version": {
            "title": "Validator Version",
            "type": "string"
          }
        },
        "required": [
          "retrieval_profile_id",
          "status",
          "validator_version",
          "candidates_found"
        ],
        "title": "RetrievalValidationSchema",
        "type": "object"
      },
...
      "SearchRetrievalSchema": {
        "additionalProperties": false,
        "description": "Run one retrieval search against one retrieval profile.",
        "properties": {
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "retrieval_profile_id": {
            "minLength": 1,
            "title": "Retrieval Profile Id",
            "type": "string"
          },
          "top_k": {
            "default": 10,
            "maximum": 25.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "retrieval_profile_id",
          "query"
        ],
        "title": "SearchRetrievalSchema",
        "type": "object"
      },
...
      "ValidateRetrievalSchema": {
        "additionalProperties": false,
        "description": "Run a controlled smoke validation against one retrieval profile.",
        "properties": {
          "retrieval_profile_id": {
            "minLength": 1,
            "title": "Retrieval Profile Id",
            "type": "string"
          }
        },
        "required": [
          "retrieval_profile_id"
        ],
        "title": "ValidateRetrievalSchema",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Chatbot SST Pipeline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/retrieval/profiles": {
      "get": {
        "operationId": "list_profiles_api_retrieval_profiles_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRetrievalProfilesSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Profiles",
        "tags": [
          "retrieval"
        ]
      },
      "post": {
        "operationId": "create_profile_api_retrieval_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRetrievalProfileSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalProfileSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Profile",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/api/retrieval/profiles/{retrieval_profile_id}/activate": {
      "post": {
        "operationId": "activate_profile_api_retrieval_profiles__retrieval_profile_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "retrieval_profile_id",
            "required": true,
            "schema": {
              "title": "Retrieval Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalProfileSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Activate Profile",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/api/retrieval/profiles/{retrieval_profile_id}/status": {
      "get": {
        "operationId": "get_profile_status_api_retrieval_profiles__retrieval_profile_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "retrieval_profile_id",
            "required": true,
            "schema": {
              "title": "Retrieval Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalProfileStatusSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Profile Status",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/api/retrieval/search": {
      "post": {
        "operationId": "search_retrieval_api_retrieval_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRetrievalSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalSearchResponseSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Search Retrieval",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/api/retrieval/validate": {
      "post": {
        "operationId": "validate_retrieval_api_retrieval_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRetrievalSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalValidationSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Validate Retrieval",
        "tags": [
          "retrieval"
        ]
      }
    }
  }
}